Move RsFormatMacroAnnotator to a separate module
